LayoutTests/platform/ios-simulator/editing/deleting/delete-at-paragraph-boundaries-001-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: shouldEndEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification


Tests: 

Deleting when a selection starts in a blank line created by a block with a BR placeholder in it and extends to the end of a block.


Expected Results:

Should see the three lines in the red box. First line should be "A". Next two lines should be empty. Insertion point should
be blinking on the second line (the first blank one).


| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <p>
|     style="margin-top: 0; margin-bottom: 0"
|     "A"
|   "
"
|   <p>
|     style="margin-top: 0; margin-bottom: 0"
|     <br>
|       class="webkit-block-placeholder"
|   "
"
|   <p>
|     style="margin-top: 0; margin-bottom: 0"
|     "A"
|   "
"
|   <p>
|     style="margin-top: 0; margin-bottom: 0"
|     <br>
|       class="webkit-block-placeholder"
|   "
"
| "
"
